Add memory64 support to wasm-smith
This commit adds support to wasm-smith to generate modules that use
memory64, namely 64-bit memories and all indices passed to memory-using
operations use 64-bit indexes instead of 32-bit indicies. The main
changes to support this were in the `code_builder.rs` file around
generating memory indices to work with and such.

This commit also refactors the fuzzers a bit as well as `SwarmConfig`.
The fuzzers now enable all features in `SwarmConfig`, even those that
are disabled by default in `SwarmConfig` due to not being implemented in
many locations (such as `memory64`). The goal here is that the fuzzing
of this repository specifically should fuzz everything supported by
wasm-smith, not just the default items.