wasm-smith: Fix calculation of `memory_offset` for 32-bit memories
This fixes an issue from #304 where maximally sized memories could panic
in `int_in_range` due to an empty range being passed in. The bug was
that the previous arithmetic in `u32`-land was now happening in
`u64`-land which means that some truncation from 4gb to 4gb-1 needed to
happen.