Stop being byte-for-byte compatible with wabt (#306)
This commit removes all reference testing against wabt in the
`roundtrip.rs` suite. Historically this project has attempted to be
byte-for-byte compatible with upstream `wabt`, but over time it's
accrued so many workarounds in both the implementation and while testing
that I don't think that it's worth it any more.

The original motivation for this change is that this module does not
appropriately roundtrip in the Rust tooling:

    (module
      (elem funcref))

The reason for this is that a hack in the implementation was changing
the `funcref` representation to `func`, which means it's round-trip as:

    (module
      (elem func))

and this was found by the fuzzers last night when reference types was
enabled for round-tripping for the first time.

Overall I don't think we have much to gain any more for byte-for-byte
compatibility with wabt. There's a number of long-open issues which
require workarounds in this crate. I think at this point this tooling is
solid enough that comparisons can be made but it isn't necessary to do
so in an automated fashion. This should make it much easier to implement
new features and such because there doesn't need to be so many
workarounds in the test suite for comparing against wabt.