Disable module linking in round-trip fuzz test (#310)
The module linking text format is well-known to not round-trip well
right now due to ambiguities with the original text format. This is
expected with the next iteration of the text format where adapter
modules are entirely separate from core wasm modules, which means that
the text format will be much more easily round-trip-able.