Don't have wasm-smith be entirely generic (#313)
Working on fuzzers using wasm-smith takes forever because they take so
long to compile. I believe the primary reason for this is that the
entire wasm-smith crate is monomorphized into each individual fuzzer due
to the usage of generics on the `Config` trait. This means each fuzzer
instantiates this entire crate, and the wasm-smith crate is no small
chunk of change.

This commit moves the configuration to being a trait object instead of a
generic, which should not lose that much in runtime and should allow
fuzzers to benefit from separate compilation of wasm-smith itself. If
this becomes a performance bottleneck for whatever reason we can also
remove the trait and simply have struct-based configuration too.