Favor generating smaller limits for tables/memories
This commit updates wasm-smith to generate, on average, smaller limits
for tables/memories than it does today. Today it uses the configuration
of the module to figure out maximums, but then it simply uses
`int_in_range` to determine the miminum/maximum. This quite frequently
generates modules that need gigabytes of memory (or multiple with
multi-memory enabled) and huge amounts of table elements. The problem is
compounded with multiple memories/tables as well.

The goal of this commit is to generate modules that require so much
memory less of the time. This commit implements a strategy where the
minimum size of a table/memory is still selected within the specified
bounds of the configuration but it's far more likely to select a size
close to the start of the range instead of the end. This should
generate tables/memories with much smaller minimum sizes much more
frequently.

In a local test ~50% of the modules generated by wasm-smith prior to
this change required more than 1GB to instantiate, and now after this
commit that's less than 1%. This is tunable as a compile-time constant
for now but if necessary we can change this to a runtime configuration
value as well.