Generate in-bounds segments much more frequently
This commit follows in the footsteps of the previous commit to make it
much more likely that wasm-smith generates a module which can actually
be instantiated. Before this commit ~40% of modules would immediately
trap without actually running any wasm code, primarily due to the fact
that element and data segments were being placed out-of-bounds in a
memory/table. The reason for this is that the offset of segments was
chosen simply with `int_in_range`, which is just as likely to generate 2
billion as it is 0.

The fix here is to employ a scheme similar to the previous commit where
the code is restructured to have a compile-time constant controlling how
likely it is that items are generated in-bounds. This means that it's
much more likely to generate a segment that's in-bounds and correctly
instantiates.

After this commit the trap rate from out-of-bounds element/data segments
drops to ~1% for generated modules.