wasm-smith: Generate fail-to-instantiate modules far less frequently (#314)
* Favor generating smaller limits for tables/memories

This commit updates wasm-smith to generate, on average, smaller limits
for tables/memories than it does today. Today it uses the configuration
of the module to figure out maximums, but then it simply uses
`int_in_range` to determine the miminum/maximum. This quite frequently
generates modules that need gigabytes of memory (or multiple with
multi-memory enabled) and huge amounts of table elements. The problem is
compounded with multiple memories/tables as well.

The goal of this commit is to generate modules that require so much
memory less of the time. This commit implements a strategy where the
minimum size of a table/memory is still selected within the specified
bounds of the configuration but it's far more likely to select a size
close to the start of the range instead of the end. This should
generate tables/memories with much smaller minimum sizes much more
frequently.

In a local test ~50% of the modules generated by wasm-smith prior to
this change required more than 1GB to instantiate, and now after this
commit that's less than 1%. This is tunable as a compile-time constant
for now but if necessary we can change this to a runtime configuration
value as well.

* Generate in-bounds segments much more frequently

This commit follows in the footsteps of the previous commit to make it
much more likely that wasm-smith generates a module which can actually
be instantiated. Before this commit ~40% of modules would immediately
trap without actually running any wasm code, primarily due to the fact
that element and data segments were being placed out-of-bounds in a
memory/table. The reason for this is that the offset of segments was
chosen simply with `int_in_range`, which is just as likely to generate 2
billion as it is 0.

The fix here is to employ a scheme similar to the previous commit where
the code is restructured to have a compile-time constant controlling how
likely it is that items are generated in-bounds. This means that it's
much more likely to generate a segment that's in-bounds and correctly
instantiates.

After this commit the trap rate from out-of-bounds element/data segments
drops to ~1% for generated modules.

* Add a script/test for testing failed instantiations

With `cargo test` this will assert that the failure rate for
instantiations is under 10%, and otherwise when running the script it'll
just run infinitely printing out how many instantiations are failing.

* Implement a more intelligent `gradually_grow` metric

Based of the cubic idea, but tweaked to have far more values inbounds
than cubic gives us by default.

* More sanity asserts in `map_custom`