Exclude canonicalize-nan locals from codegen
This commit updates wasm-smith to exclude injected locals from the nan
canonicalization insertion from being candidates for generated
`local.{get,set,tee}` instructions. The injected locals contain
non-canonical NaN values which, if used, defeats the purpose of
canonicalizing NaN since it leaks the non-canonical value into the rest
of the program.

The fix implemented here is to accumulate the extra types in a separate
vector that's appened to the normal `locals` vector at the end of code
generation.