Fix a stack overflow parsing instance/module types
This commit fixes the parsing of instance/module types to have a fixed
recursion limit after which it's an error to keep parsing. Currently
there's no restriction on recursion for these types (part of the module
linking proposal), so these are given the same treatment as nested
modules where it's not easily feasible to fix the recursion here like it
is with expressions.