Add support for the extended-consts wasm proposal
This commit updates the upstream testsuite that we test against and at
the same time implements a new proposal added to the suite, the
extended-const proposal. This proposal adds a few more instructions that
are valid in an init-expr context such as global initialization
expressions.

The internal changes to implement this new proposal are to use an
`OperatorsValidator` in its full glory when validating a global or
element initialization expression. This means that we get all the new
operators supported for free, and an allow-list is checked beforehand
for init-expressions to ensure only valid "const" instructions are
allowed.

This also necessitated some tweaks to parsing and printing of
more-than-one-instruction expressions which are accounted for here as
well.