Optimize `*.wat` parsing (#389)
* Optimize `*.wat` parsing

This commit attempts to resolve at least some slowness in text parsing
for wasm. With some profiling the main source of slowdown for parsing
the text format appeared to be the actual utf-8 decoding of the source,
apparent because the source was iterated over `char`-by-`char`. This
decoding led to lots of branch-y code even for simple operations such as
skipping whitespace.

The change in this commit is to change the lexer of the text format to
working with `&[u8]` internally instead of `CharIndices`. This means
that the lexer will utf-8 decode as little as possible from now on and
only reading `char` values as necessary for lexing operations such as
strings. Otherwise the text format is primarily ASCII and doesn't need
to go into the realm of utf-8 decoding. Also note that the lexer still
operates over `&str` at the API-level since wasm source text is required
to be valid utf-8.

For parsing a wasm-smith-generated-module this cause the entire
`wat2wasm` operation to be 40% faster. The vast majority of the
remaining time is still spent in lexing, almost exclusively in skipping
whitespace (since heavily nested blocks lead to lots of whitespace).
Unfortunately I don't really know of an easy way to optimize skipping of
whitespace, but at least the problem is quite tractable where the goal
is to find the first-non-whitespace-byte in a list of bytes, so this
should be ripe for optimization in the future.

* Review comments

* Tidy up a bit