Protect against printing some invalid modules (#395)
This commit fixes a fuzz bug found on oss-fuzz where printing some input
would take a pathologically long time. The input was an invalid wasm
module which had *just* the right shape to trick wasmprinter into
printing the code section thousands of times where the code section
happened to have thousands of locals, causing a huge amount of time to
be spent printing locals. This is fixed with two different measures:

* The code section is now printed at most once
* The code section is asserted to have the same number of entries as the
  function section.

The hope is that this gets printing back on track for invalid modules to
not so easily take up extra compute resources for small modules.