wasm-mutate: Refactor and simplify dfg module
I was curious to add support for some simd instructions and noticed that
the dfg module could do from some simplifications. This should make it a
bit easier to add new instructions in the future hopefully. This commit
is intended to be a pure refactoring and not actually change anything
(e.g. no simd support added here).