wasm-mutate: Simplify the `build_expr_innner` function
This commit aims, like #425, to make adding new `Lang` variants easier.
To that end the `rebuild::build_expr_inner` function was doing a match
and handling each case of `Lang` separately, but instead all it
internally needed to do was to update each of the operands of each
variant. This commit switches the function to using
`Language::children_mut` which removes the need for a match over all
`Lang` variants and it's naturally handled by the `children_mut`
accessor which can update in-place.