wasm-mutate: Simplify push-operand in expr2wasm
Further along the lines of "make it easier to add new `Lang` nodes" this
commit simplifies the operation of pushing child nodes from `Lang` onto
the operand worklist as part of the expr2wasm pass. Previously children
were manually iterated over in a `match` and this instead switches it to
using the `Language::children` accessor.

This wasn't a trivial change because `Select` needed to change to match
the order of operands that wasm expects. Previously `Select` was
expressed as `(select condition if-true if-false)` and now it's updated
to match wasm where it's `(select if-true if-false condition)`. Once
that was updated however, along with the relevant rules, everything
appeared ship-shape.