Update some exception-handling error messages
This updates the upstream spec test suite which has a number of changes
to the exception-handling proposal. I find some of the test expectations
to be quite odd so I've gone ahead and ignored a number of new tests and
they can get un-ignored as the proposal moves further along and the
tests are formalized as either "yes this is what to do" or "no that was
a mistake".