Updated upstream spec test suite (#432)
* Tweak multi-memory index parsing

* Allow non-`(memory ...)` indices in simd lane-related stores/loads.
* Swap the text format operands for `memory.init` to match `table.init`

Overall this brings parsing in-line with upstream spec-tests are
expecting. The parsing of the lane-related logic is sort a funky but not
necessarily overly-funky.

* Update some exception-handling error messages

This updates the upstream spec test suite which has a number of changes
to the exception-handling proposal. I find some of the test expectations
to be quite odd so I've gone ahead and ignored a number of new tests and
they can get un-ignored as the proposal moves further along and the
tests are formalized as either "yes this is what to do" or "no that was
a mistake".