Further simplify `Language for Lang`
... with macro soup.

This commit is a further attempt to greatly simplify the implementation
of `Language for Lang`. The goal here is to use a `macro_rules!` based
macro to easily defined the various `egg::Language` operations for us.

While the goal of this commit is achieved, one should always ask "but
what is the cost". Sometimes we have GPS and yet we're still lost. The
`lang!` macro introduced as part of this crate is quite gnarly and is
not easy to read or write. That being said the hope is that this macro
will very rarely change, if ever, instead simply making it easier to add
new variants to the `enum Lang` with a little metadata to automatically
get string parsing, string printing, and children accesses.

To achieve this the `Lang` enum is now refactored and categorized into
a few styles of variants:

* Empty variants, like before, have no children and no payload.
* Some variants have both children and a payload. The payload (such as
  the call index of a function) is the first type and the children list
  is the second type. Variants like `I32Load` were refactored to have a
  `MemArg` as the first type which encapsulates all static information.
* Some variants only have children, and that's expressed as their single
  type.
* Finally some variants have no children, but do have a payload, and
  they're also expressed with a single type.

The macro does some special-casing on the structure of a variant and
whether or not the first argument looks like a "list of children"
argument to implement various methods. Some special-cases may need to be
added in the future if necessary but ideally shouldn't be too too hard.

Otherwise the other major change from this commit is to the text format
itself. To facilitate the same pattern of parsing for everything the
`0_i32` terms are now parsed as `i32.const.0`.

Finally I've tried to document the macro the best I can, but in my
opinion there's no real way to get away from the macro soup that it is.