wasm-mutate: Add bulk memory support to DFG
This commit adds support to the wasm-mutate DFG for the bulk-memory
proposal, notable table/memory copying instructions as well as table
operation instructions. No rules are added yet for rewriting, for
example simply removing `elem.drop` or `data.drop`, but I'm hoping that
will be easier to add later with the DFG nodes already added.