Rename `wasmparser::TypeOrFuncType` to `BlockType`.
This commit renames the confusing `TypeOrFuncType` type to `BlockType` to
better represent what is actually being described.

As a result, this commit also removes the "empty" variant from the core `Type`
enumeration as this was bleeding in from what should have been an "empty"
variant on `BlockType` originally.

It also removes `ExnRef` and `Func` variants as they were removed from the
related proposals a long time ago now.