Reimplement module adapter and component encoding in `wasm-encoder`.
This commit reimplements the module adapter and component encoding in
`wasm-encoder` to unify the various section types.

The support for module adapters now supersedes the previous support for module
linking in `wasm-encoder`.

The `adapter` and `component` submodules have been removed and the duplicate
section types eliminated.

Certain section types that encode with different formats now take a
`SectionEncodingFormat` enumeration on construction to instruct what format to
use for the encoding.