Update for latest component model binary AST changes.
This commit updates the encoder to use the latest version of the component
model binary AST.

This does not remove the `ComponentSection` trait for now as there still
remains substantial differences between the component and module ASTs.

However, the recent changes do make things a bit clearer in terms of what comes
from a core wasm module and what comes from a component.

To that effect, certain sections that share a name (e.g. `types`) have had
their encoders split in two (e.g. a `TypeSection` encoder for modules and a
`ComponentTypeSection` for components).

The code duplicated between these two types is quite minimal and having two
types allows for an API surface that has some basic correctness to it (i.e.
it's no longer possible to encode a module type section type in a component
type section).

We'll revise this design as the component model proposal evolves.