wasm-mutate: avoid mutating empty code sections (#466)
The `Rng::gen_range`, which both `SnipMutator` and
`FunctionBodyUnreachable` mutators utilize to pick a function to mutate,
has a panicking failure mode when the range provided to it is empty.

This failure mode may be triggered if the mutation ends up removing all
of the functions, thus leaving an empty code section.

One could argue that removing the section as it becomes empty would be a
cleaner fix for this particular problem. Being able to verify operation
of the predicate when fed modules with empty sections is an important
feature that should be preserved.