Disable previous module linking tests.
This commit renames a bunch of wat tests to have a "fixme" extension. Once the
`wast` crate is updated such that the component model is supported, these tests
will need to be updated and renamed back to the `.wat` extension.

The previous module linking wat tests were renamed to `component-model` and are
explicitly skipped currently. As above, these tests will need to be updated and
the exclusion removed once the `wast` crate supports the component model.

A few other tests were fixed up as a part of this commit.