Check limits once for most sections.
This commit changes the `ensure_module_section` and `ensure_component_section`
functions to provide information for limit checks so that the limit check
doesn't have to be inside of a loop for the majority of sections.

Import limits have been removed since imports check the individual limits of
the item being imported each iterator.

Similarly, aliases in components don't have their own limits but instead check
the limits for each item being aliased.