Implement `into` option validation for the component model.
This commit implements `into` option validation for the component model.

It checks to see if the into option is required based on the interface
signature of the function being lifted or lowered.

Additionally, it ensures that the into option specifies a value module instance
index that exports the expected canonical ABI items.