Implement wit2wasm and wasm2wit tools.
This commit implements simple CLI tools for encoding a single wit file to a
WebAssembly component (`wit2wasm`) or decoding a WebAssembly component to a wit
file (`wasm2wit`).

The latter tool right now only works with the output of `wit2wasm` and may
eventually be made more generic to spit out a wit representation for
all imported/exported interfaces of a WebAssembly component.