wasm-smith: Generate module types with canonical ABI functions
It is highly unlikely that we would ever generate module types that export the
canonical ABI's expected `realloc` and `free` functions by chance, and passing
certain types across an interface types boundary requires these functions, so
special case them and generate them most of the time for most modules. Do the
same for the `memory` export.