Clamp the maximum nested printing level in wasmprinter
Occasionally I've seen timeouts on oss-fuzz related to printing
small-ish wasm modules that have very deeply nested blocks. Each level
of nesting increases the size of subsequent lines for later blocks,
meaning that it's pretty easy to generate a small wasm module (<1M)
which generates hundreds of megabytes of textual output (>200M). This
limit is intended to be a small stopgap against that by clamping the
maximum nesting level to a reasonable-ish value that ideally shouldn't
be hit in practice but may still show up from time to time. This doesn't
actually affect the interpretation of the textual output, only the
representation.