fuzz: Enable more validation feature in the mutate fuzzer
It's possible for wasm-mutate to take a module without simd enabled, for
example, and add simd features into it such as a type with `v128` or
similar. This means that the features for validation cannot be based
entirely on the input but rather some features need to be
unconditionally enabled (those that `wasm-mutate` could add at any
time).