Add some basic round-trip tests for wit <-> wasm.
This commit adds some tests to round trip from wit files to a component
(in-memory), decoding the component bytes back to an `Interface`, and then
printing the interface to compare against the original input file (which is
expected to be from a previously "blessed" test run).

Also fixes the printing of flags to use `flags` rather than record of bools in
`InterfacePrinter` and functions with the same signature decoding to the wrong
names.