wasmparser: return type information built up from validation.
This commit changes `Validator::End` to return type information for the module
or component that was just validated.

The type information will help other tools so that they don't have to duplicate
what the validator is doing to determine the types of things in a module or
component's index spaces.