Use `IndexMap` and `IndexSet` to get stable associative collections.
For implementing the `InterfaceType` enumeration in the validation types, we
need an associative collection for fast lookups when doing subtype checks.

Additionally, for consumers of the validator's type information, we need a
stable insertion-order-based iteration of the entries (e.g. for things like
enumerations which are order dependent).

This commit replaces the use of `HashMap` and `HashSet` with `IndexMap` and
`IndexSet` from the `indexmap` crate.