wasmparser: return type information built up from validation.  (#503)
* Move operators validator into validator module.

Just a simple refactoring of moving the `OperatorValidator` to be under the
`validator` module rather than having its own.

* wasmparser: return type information built up from validation.

This commit changes `Validator::End` to return type information for the module
or component that was just validated.

The type information will help other tools so that they don't have to duplicate
what the validator is doing to determine the types of things in a module or
component's index spaces.

* wasmparser: fix `Validator::validate_payload` for nested parsers.

This commit fixes `Validator::validate_payload` to correctly return back the
nested parsers when parsing a nested module or component.

* wasmparser: add a test for module type information in the validator.

This commit implements a simple unit test to ensure that module type
information is returned correctly when validating a core module.

* Fix fuzz target compilation.

* Use `IndexMap` and `IndexSet` to get stable associative collections.

For implementing the `InterfaceType` enumeration in the validation types, we
need an associative collection for fast lookups when doing subtype checks.

Additionally, for consumers of the validator's type information, we need a
stable insertion-order-based iteration of the entries (e.g. for things like
enumerations which are order dependent).

This commit replaces the use of `HashMap` and `HashSet` with `IndexMap` and
`IndexSet` from the `indexmap` crate.

* wasmparser: add ability to get a type index's type id.

This adds the `id_from_type_index` method to `Types` so that users of the
validator's type information can figure out what the "type id" would be for
types in the type index space.