wasmparser: remove enforcement that `into` instances must export a memory. (#506)
Currently, the component model validator requires that an instance pointed at
by the `into` canonical option export a memory named `memory`.

However, we have a chicken-and-the-egg problem when lowering imported
functions: the functions' `into` options must point at an instance, but we
don't have the instance yet because we're lowering the functions that the inner
core module will be importing.

To work around this problem, a small "shim" instance is used that has a
`funcref` table pointing at the canonical ABI functions expected by the `into`
option. The table gets populated after the inner core module gets instantiated.

Thus the `into` options for lowering imported functions point at this shim
instance rather than the actual core module's instance for the component.

As the shim instance can't export the same memory as the actual core module,
this validation enforcement prevents this from working.