Add initial memory64 and multi-memory support to wasm-mutate
This fixes at least one issue I was encountering on #511 where I forgot
that the modules were using memory64 which meant that the
`allow_memarg64` method needed to be enabled to enable parsing 64-bit
offsets in memargs. This also touches up a few other locations to
support memory64 and multi-memory, enabling fuzzing as well. This gets
the exact test case on #511 to reduce much further in the shape of the
outer module (e.g. elements, imports, etc all removed), although the
function itself causing the issue is still not being reduced.