wasm-shrink: Avoid local minima
This commit allows individual mutations to grow the code size of the Wasm that
we are shrinking with a low probability. At the same time, we decouple the
concept of the smallest interesting Wasm we've discovered so far and the current
Wasm that we are mutating and trying to find more smaller interesting Wasms
from. This technique is borrowed from [MCMC]. As an example, it allows us to
accept an intermediate mutation that replaces a `func.ref $f` with a
`ref.null` (which by itself does not shrink code size) so that a later mutation
can recognize that `$f` might be dead code now and can be completely
removed (which is generally a large code size win).

Additionally, this refactors how the `WasmShrink::run` method was defined and
factors its local variable state out into a `ShrinkRun` struct and the closures
that were defined within it into methods on `ShrinkRun`.

[MCMC]: https://en.wikipedia.org/wiki/Markov_chain_Monte_Carlo