`wasmprinter`: place limit of number of functions in a module.
`wasmprinter` was updated to reserve the code vector based on the number of
functions in the module.

Previously the reservation was unlimited, meaning a fuzzed or maliciously
crafted module with an exceptionally large number of functions would force the
printer to make a very large allocation.

This commit places the same limit we use in `wasmparser`: the module cannot
have more than 1 million functions.