Update the spec test suite
This updates the spec test suite submodule. At the same time I realized
that we weren't actually running `assert_malformed` tests for
binary-encoded modules, so I enabled that. Naturally the error messages
between the spec interpreter and wasmparser had lots of differences, so
I updated a number of error messages in wasmparser to either align with
the spec interpreter or have more context as well (some errors are more
detailed now).

Overall though I did end up still "skipping" a lot of tests by making
error-equality pretty loose in a bunch of situations. The major
difference between the spec interpreter and wasmparser is that the spec
interpreter will read past section boundaries to decode items, whereas
wasmparser will not. This leads to quite a few test cases which are
actually somewhat broken and their length indicators for sections don't
actually encompass the entire section. This means that the error popping
out of the spec interpreter is valid (since it presumably validations
setion lengths later) but the errors from wasmparser are much different.
I don't think this is a huge issue because so long as everything is
rejected for somethat along the lines of the same reason we should be
good to go.