Remove wabt submodule
This is a revival of #357 where I'm attempting to remove the wabt
submodule from the wasm-tools crate. This submodule takes awhile to
checkout for git dependencies and also building wabt seems to add a
nontrivial chunk of time on the Windows test builder otherwise as well.
We continue to not really need anything from wabt any more so this
cleans up some code while I'm at it for bits and pieces that reference
wabt.