Handle situations where extant entities exceed min
In some instances it is possible that before imports or types are
generated that there will already be more of the thing than the
specified minimum. `saturating_sub` here works to clamp these values
down to 0. I ended up also applying the same treatment to `max_imports`
as well, if we somehow end up with more of the entity than specified by
the `max` configuration.