wasm-mutate: remove empty sections
I have found that `wasm-shrink` would reduce the reproducer case to 1 or
two relevant sections and leave a bunch of other empty sections laying
around. This change introduces a mutation to `wasm-mutate` which is able
to clean-up these sections. Since the removed sections are empty,
removing them should have no impact on the validity of the module.