wasmparser: Factor out and expose a `CustomSectionReader`
This allows users who have a slice of section data, but not the preceding
section ID, to parse the name and data for a custom section when they happen to
know that the section data is for a custom section. This matches the use of
other section readers, for example if you know you have an element section, you
can use `ElementSectionReader::new` to read that section (without the preceding
element section id).