wasmprinter: Always print the index of item definitions
I often find myself with the need to find function N in the text format
of a wasm file but the default output of `wasmprinter` doesn't actually
print out indices if items are named. For example instead of:

    (func $foo)

this will now print out

    (func $foo (;2;))

which can be quite helpful if you're otherwise looking for function 800
in a 2000 function file.