wasmparser: Fix a bug with validating nested components
Without this change, all nested components would incorrectly be reported as
invalid due to an off-by-one error. We need to add the current component to the
parent if `self.components.len() >= 1` but were checking `self.components.len()
> 1` instead. Maybe we were accidentally including the current component in the
count? But that was popped off `self.components` a few lines of code up. Either
way, using `if let Some(...) = self.components.last_mut()` the same way we do
for nested modules fixes this bug and also avoids a follow up call to `unwrap`.