wasm-smith: Avoid adding more types than configured when lifting core functions
When we lift core functions, we try to reuse already defined interface function
types for the lifting, but if there isn't one already defined, then we will
define it on the spot. However, we weren't previously checking whether this
would cause us to go over the configured maximum number of types. We could add
too many types and then we could cause subtraction overflows in other code that
assumes that the number of types is always less than the configured maximum.

The fix is to check whether we have budget to add new types before doing so
during core Wasm function lifting.